Assignment 8
The Assignment 8 class displays a menu of choices to a user and performs
the chosen task. It will keep asking a user to enter the next choice
until the choice of 'Q' (Quit) is entered. It will create an object of
Store containing a list of computers and customers, and perform
insertion, deletion, search, sorting, displaying, closing,
serialization, deserialization, writing to a file, and reading from a
file.